% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{cal_rank_score}
\alias{cal_rank_score}
\title{calculate sku rank}
\usage{
cal_rank_score(df, ..., amount_weight = 0.2, quantity_weight = 0.8)
}
\arguments{
\item{df}{a tbl or dataframe}

\item{...}{Variables to group by. All tbls accept variable names.
Some tbls will accept functions of variables. Duplicated groups
will be silently dropped.}

\item{amount_weight}{defaults to 0.2}

\item{quantity_weight}{defaults to 0.8}
}
\description{
calculate sku rank
}
