% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{calc_del_rate}
\alias{calc_del_rate}
\title{calculate the deletion rate for deleted order data}
\usage{
calc_del_rate(df)
}
\arguments{
\item{df}{a tbl or dataframe}
}
\description{
calculate the deletion rate for deleted order data
}
