% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{common_summary_table}
\alias{common_summary_table}
\title{summary table of mesures of location}
\usage{
common_summary_table(df, var, digits = 2, caption = NULL)
}
\arguments{
\item{df}{dataframe}

\item{var}{variable that you compute measures for}

\item{digits}{the maximum number of digits for numeric columns (passed to
\code{round()}); it can also be a vector of length \code{ncol(x)} to set
the number of digits for individual columns}

\item{caption}{the table caption}
}
\description{
summary table of mesures of location
}
