% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{connect_biee}
\alias{connect_biee}
\title{Connect to biee using simon's username and password}
\usage{
connect_biee(host, port, user, password)
}
\arguments{
\item{host}{host name; Usually the ip address of the host}

\item{port}{the port number;}

\item{user}{the user name}

\item{password}{password}
}
\description{
Connect to biee using simon's username and password
}
