% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{connect_postgres}
\alias{connect_postgres}
\title{connect to postgres database}
\usage{
connect_postgres(user = "postgres", password = "mysecretpassword",
  dbname = "postgres", host = "127.0.0.1", port = 5432)
}
\arguments{
\item{user}{the user name}

\item{password}{password}

\item{dbname}{the name of the database on the host, or the database file name}

\item{host}{host name; defaults to 127.0.0.1}

\item{port}{the port number; defaults to 5432}
}
\description{
connect to postgres database
}
