% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{connect_sql_server}
\alias{connect_sql_server}
\title{Connect SQL Server}
\usage{
connect_sql_server(server = "127.0.0.1", port = 1433, database = "Master",
  useNTLMv2 = "false", user, password)
}
\arguments{
\item{server}{server address; defaults to 127.0.0.1}

\item{port}{port number}

\item{database}{the database name. Defaults to Master}

\item{useNTLMv2}{a authentication method. Defaults to false}

\item{user}{user name}

\item{password}{the password}
}
\description{
Connect SQL Server
}
