% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{create_workbook_sku_image}
\alias{create_workbook_sku_image}
\title{Create a workbook for dataframe with sku image}
\usage{
create_workbook_sku_image(df, df_image_title_col_name, df_image_url_col_name,
  image_col_name = "Image", widths = 27, heights = 144,
  image_file_extension = ".JPG", width = 5, height = 5, units = "cm",
  dpi = 300, image_path = "~/images")
}
\arguments{
\item{df}{a tbl or dataframe}

\item{df_image_title_col_name}{the colname of the image title (usually the sku)}

\item{df_image_url_col_name}{the colname of image url column}

\item{image_col_name}{the colname for the image in df}

\item{widths}{widths to set cols to specified in Excel column width units or "auto" for automatic sizing. The widths argument is
recycled to the length of cols.}

\item{heights}{Heights to set rows to specified in Excel column height units.}

\item{image_file_extension}{the file extension of the image}

\item{width}{Width of figure.}

\item{height}{Height of figure.}

\item{units}{Units of width and height. Can be "in", "cm" or "px"}

\item{dpi}{Image resolution used for conversion between units.}

\item{image_path}{where the image locates; default to ~/images}
}
\value{
an openxlsx workbook
}
\description{
Create a workbook for dataframe with sku image
}
