% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{datatable_cn}
\alias{datatable_cn}
\title{Create an HTML table widget using the DataTables library with Chinese menu}
\usage{
datatable_cn(data, caption = NULL, rownames = FALSE, colnames,
  extensions = "Buttons", pageLength = pageLength, lengthMenu = c(5,
  10^(1:5)), buttons = c("copy", "csv", "excel", "pdf", "print"))
}
\arguments{
\item{data}{a data object (either a matrix or a data frame)}

\item{caption}{the table caption; a character vector or a tag object
generated from \code{htmltools::tags$caption()}}

\item{rownames}{\code{TRUE} (show row names) or \code{FALSE} (hide row names)
or a character vector of row names; by default, the row names are displayed
in the first column of the table if exist (not \code{NULL})}

\item{colnames}{if missing, the column names of the data; otherwise it can be
an unnamed character vector of names you want to show in the table header
instead of the default data column names; alternatively, you can provide a
\emph{named} numeric or character vector of the form \code{'newName1' = i1,
'newName2' = i2} or \code{c('newName1' = 'oldName1', 'newName2' =
'oldName2', ...)}, where \code{newName} is the new name you want to show in
the table, and \code{i} or \code{oldName} is the index of the current
column name}

\item{extensions}{a character vector of the names of the DataTables
extensions (\url{https://datatables.net/extensions/index})}

\item{pageLength}{the length of page}

\item{lengthMenu}{the menu of length}

\item{buttons}{buttons shows}
}
\description{
Create an HTML table widget using the DataTables library with Chinese menu
}
