% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{days_df}
\alias{days_df}
\title{Generate a tbl of day week month quarter semester year.
Mainly used for joining data}
\usage{
days_df(sep = "~", with_year = TRUE)
}
\arguments{
\item{sep}{a character string to separate the terms.  Not
    \code{\link{NA_character_}}.}

\item{with_year}{logical indicating whether or not to include the quarter's
year.}
}
\description{
Generate a tbl of day week month quarter semester year.
Mainly used for joining data
}
