% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{download_image_url}
\alias{download_image_url}
\title{download image of the image column}
\usage{
download_image_url(df, df_image_title_col_name, df_image_url_col_name,
  lsp_label = "LSP", image_file_extension = ".JPG",
  image_path = "~/images")
}
\arguments{
\item{df}{a tbl or dataframe}

\item{df_image_title_col_name}{the colname of the image title (usually the sku)}

\item{df_image_url_col_name}{the colname of image url column}

\item{lsp_label}{the common label of lsp (linshi product); There is no image for lsp}

\item{image_file_extension}{the file extension of the image}

\item{image_path}{where the image locates; default to ~/images}
}
\value{
df
}
\description{
download image of the image column
}
