% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{downloadfile}
\alias{downloadfile}
\title{shiny download module Server}
\usage{
downloadfile(input, output, session, x, filename)
}
\arguments{
\item{input}{refer to \url{https://shiny.rstudio.com/articles/modules.html}}

\item{output}{refer to \url{https://shiny.rstudio.com/articles/modules.html}}

\item{session}{refer to \url{https://shiny.rstudio.com/articles/modules.html}}

\item{x}{data frame to write to disk}

\item{filename}{A string of the filename, including extension, that the
user's web browser should default to when downloading the file; or a
function that returns such a string. (Reactive values and functions may be
used from this function.)}
}
\description{
shiny download module Server
}
