% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{downloadfileOutput}
\alias{downloadfileOutput}
\title{shiny download module UI}
\usage{
downloadfileOutput(namespace, label)
}
\arguments{
\item{namespace}{The character vector to use for the namespace. This can have
any length, though a single element is most common. Length 0 will cause the
\code{id} to be returned without a namespace, and length 2 will be
interpreted as multiple namespaces, in increasing order of specificity
(i.e. starting with the top-level namespace).}

\item{label}{The label that should appear on the button.}
}
\description{
refer to https://shiny.rstudio.com/articles/modules.html
}
