% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biee.R
\name{fetch_biee_sql}
\alias{fetch_biee_sql}
\title{Fetch data from database}
\usage{
fetch_biee_sql(conn, con, ...)
}
\arguments{
\item{conn}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[=dbConnect]{dbConnect()}}.}

\item{con}{a \link{connection} object or a character string.}

\item{...}{user defined variables and corresponding parameters in sql statements}
}
\description{
Fetch data from database
}
