% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis.R
\name{fetch_data}
\alias{fetch_data}
\title{Generate a function to fetch data from cache or get new data with a query if there is no cache}
\usage{
fetch_data(exists_f, get_f, set_f)
}
\arguments{
\item{exists_f}{function used to check if cache exists.}

\item{get_f}{function used to get data from cache}

\item{set_f}{function used to set new data into cache if data not exists.}
}
\description{
Generate a function to fetch data from cache or get new data with a query if there is no cache
}
