% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{fix_image_link}
\alias{fix_image_link}
\title{Add image url prefix to image column}
\usage{
fix_image_link(df, df_image_url_col_name, image_url_prefix = NULL)
}
\arguments{
\item{df}{a tbl or dataframe}

\item{df_image_url_col_name}{the colname of image url column}

\item{image_url_prefix}{the prefix for the image url column}
}
\value{
df
}
\description{
Add image url prefix to image column
}
