% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{geom_ecdf_s}
\alias{geom_ecdf_s}
\title{ecdf geom with prefered theme}
\usage{
geom_ecdf_s(prefix = "$", suffix = "", label = "")
}
\arguments{
\item{prefix}{Symbols to display before and after value.}

\item{suffix}{Symbols to display before and after value.}

\item{label}{The title of the respective axis (for \code{xlab()} or \code{ylab()}) or
of the plot (for \code{ggtitle()}).}
}
\description{
ecdf geom with prefered theme
}
