% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{geom_fill_bar_s}
\alias{geom_fill_bar_s}
\title{filled bar geom}
\usage{
geom_fill_bar_s(prefix = "$", suffix = "", palette = "Paired",
  position = "dodge")
}
\arguments{
\item{prefix}{Symbols to display before and after value.}

\item{suffix}{Symbols to display before and after value.}

\item{palette}{If a string, will use that named palette.  If a number, will
index into the list of palettes of appropriate \code{type}}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}
}
\description{
filled bar geom
}
