% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{geom_flip_bar_s}
\alias{geom_flip_bar_s}
\title{fliped bar geom with perfered theme}
\usage{
geom_flip_bar_s(prefix = "$", suffix = "", low = "#00AFBB",
  high = "#FC4E07")
}
\arguments{
\item{prefix}{Symbols to display before and after value.}

\item{suffix}{Symbols to display before and after value.}

\item{low}{Colours for low and high ends of the gradient.}

\item{high}{Colours for low and high ends of the gradient.}
}
\description{
fliped bar geom with perfered theme
}
