% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{geom_point_line_s}
\alias{geom_point_line_s}
\title{point line geom with prefered theme}
\usage{
geom_point_line_s(prefix = "$", suffix = "", palette = "Paired", ...)
}
\arguments{
\item{prefix}{Symbols to display before and after amount.}

\item{suffix}{Symbols to display before and after amount.}

\item{palette}{If a string, will use that named palette.  If a number, will
index into the list of palettes of appropriate \code{type}}

\item{...}{Other arguments passed on to \code{\link{discrete_scale}}
to control name, limits, breaks, labels and so forth.}
}
\description{
point line geom with prefered theme
}
