% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{geom_point_smooth}
\alias{geom_point_smooth}
\title{point smooth geom with prefered theme}
\usage{
geom_point_smooth(prefix = "$", suffix = "", angle = 45)
}
\arguments{
\item{prefix}{Symbols to display before and after amount.}

\item{suffix}{Symbols to display before and after amount.}

\item{angle}{Angle (in \eqn{[0, 360]})}
}
\description{
point smooth geom with prefered theme
}
