% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{get_shiny_input_values}
\alias{get_shiny_input_values}
\title{Concatenate shiny input value into character string}
\usage{
get_shiny_input_values(input, exclude_start_chr = "table")
}
\arguments{
\item{input}{shiny input object. Always should be input}

\item{exclude_start_chr}{a character vector. If the name of input start with the character, exclude its value(s)}
}
\description{
Concatenate shiny input value into character string
}
