% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{kill_db_connections}
\alias{kill_db_connections}
\title{kill all open connections at once}
\usage{
kill_db_connections(drv = RPostgreSQL::PostgreSQL())
}
\arguments{
\item{drv}{A object inheriting from DBIDriver. Defaults to RPostgreSQL::PostgreSQL()}
}
\description{
refer to \url{https://stackoverflow.com/questions/32139596/cannot-allocate-a-new-connection-16-connections-already-opened-rmysql}
}
