% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biee.R
\name{read_biee_sql}
\alias{read_biee_sql}
\title{Read biee sql statements from file}
\usage{
read_biee_sql(con, ...)
}
\arguments{
\item{con}{a \link{connection} object or a character string.}

\item{...}{user defined variables and corresponding parameters in sql statements}
}
\value{
sql statement with user defined variables
}
\description{
Read biee sql statements from file
}
