% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis.R
\name{redis_cache_time}
\alias{redis_cache_time}
\title{Generate a function to return the cache time of a key with specified connection}
\usage{
redis_cache_time(redis_con, time_to_live)
}
\arguments{
\item{redis_con}{an interface to Redis. See \code{\link[redux]{hiredis}}}

\item{time_to_live}{time to live in seconds. \code{\link[lubridate]{dhours}} can be used.}
}
\description{
Warp TTL see \url{https://redis.io/commands/ttl}
}
\details{
Use \code{\link[pryr]{unenclose}} to check the generated function
}
\section{Closure Arguments}{

\itemize{
\item key. Redis key. A string.
}
}

