% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{send_mail}
\alias{send_mail}
\title{Send emails}
\usage{
send_mail(from = "simon@inc.tvc-tech.com", to, subject = "", body = "",
  host.name = "smtp.exmail.qq.com", port = 465, user.name, passwd,
  ssl = TRUE, authenticate = TRUE)
}
\arguments{
\item{from}{A valid email address of the sender.}

\item{to}{A character vector of recipient valid email addresses.}

\item{subject}{Subject of the email.}

\item{body}{Body of the email as text. If the parameter body refers to an existing file location, the text of the file is parsed as body of the email.}

\item{host.name}{the smtp server}

\item{port}{smtp port; defaults to 465}

\item{user.name}{usually same as the email address}

\item{passwd}{password of the email}

\item{ssl}{Whether use ssl or not; defaults to true}

\item{authenticate}{A boolean variable to indicate whether authorization is required to connect to the SMTP server. If set to true, see details on parameters required in smtp parameter.}
}
\description{
Send emails
}
