% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{summarise_del}
\alias{summarise_del}
\title{summarise the deleted order data}
\usage{
summarise_del(df, ...)
}
\arguments{
\item{df}{a tbl or dataframe}

\item{...}{For \code{enexprs()}, \code{ensyms()} and \code{enquos()}, names of
arguments to capture without evaluation (including \code{...}). For
\code{exprs()} and \code{quos()}, the expressions to capture unevaluated
(including expressions contained in \code{...}).}
}
\description{
summarise the deleted order data
}
