% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_excel_image}
\alias{write_excel_image}
\title{Write dataframe with image url to excel file with image column}
\usage{
write_excel_image(df, filename = tempfile(fileext = ".xlsx"),
  df_image_title_col_name, df_image_url_col_name, image_col_name,
  image_path = "/tmp/images")
}
\arguments{
\item{df}{a tbl or dataframe}

\item{filename}{a file name to write to}

\item{df_image_title_col_name}{the colname of the image title (usually the sku)}

\item{df_image_url_col_name}{the colname of image url column}

\item{image_col_name}{the colname for the image in df}

\item{image_path}{the path to store images locally, defaults to "/tmp/images"}
}
\description{
Write dataframe with image url to excel file with image column
}
