% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{yisoweek}
\alias{yisoweek}
\title{get the year of iso week}
\usage{
yisoweek(x)
}
\arguments{
\item{x}{a date-time object}
}
\value{
the year of the date in iso week
}
\description{
get the year of iso week
}
