% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{geom_ecdf_s}
\alias{geom_ecdf_s}
\title{ecdf geom with prefered theme}
\usage{
geom_ecdf_s(prefix = "$", suffix = "", label = "")
}
\arguments{
\item{prefix}{Symbols to display before and after amount.}

\item{suffix}{Symbols to display before and after amount.}

\item{label}{The text for the axis, plot title or caption below the plot.}
}
\description{
ecdf geom with prefered theme
}
