% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis.R
\name{redis_exists}
\alias{redis_exists}
\title{Generate a function to redis exists with specified connection}
\usage{
redis_exists(redis_con)
}
\arguments{
\item{redis_con}{an interface to Redis. See \code{\link[redux]{hiredis}}}
}
\description{
Warp EXISTS. see \url{https://redis.io/commands/exists}
}
\details{
Use \code{\link[pryr]{unenclose}} to check the generated function
}
\section{Closure Arguments}{

\itemize{
\item key. Redis key. A string.
}
}

