% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail.R
\name{send_mail}
\alias{send_mail}
\title{Send emails}
\usage{
send_mail(from, to, subject = "", body = "", encoding = "utf-8",
  host.name, port = 465, user.name, passwd, ssl = TRUE,
  authenticate = TRUE, attach.files = NULL)
}
\arguments{
\item{from}{A valid email address of the sender.}

\item{to}{A character vector of recipient valid email addresses.}

\item{subject}{Subject of the email.}

\item{body}{Body of the email as text. If the parameter body refers to an existing file location, the text of the file is parsed as body of the email.}

\item{encoding}{Character encoding to use for the email. Supported encodings include iso-8859-1 (default), utf-8, us-ascii, and koi8-r.}

\item{host.name}{the smtp server}

\item{port}{smtp port; defaults to 465}

\item{user.name}{usually same as the email address}

\item{passwd}{password of the email}

\item{ssl}{Whether use ssl or not; defaults to true}

\item{authenticate}{A boolean variable to indicate whether authorization is required to connect to the SMTP server. If set to true, see details on parameters required in smtp parameter.}

\item{attach.files}{A character vector of paths in the file system linking to files or *valid* URLs to be attached to the email (see details for more info on attaching URLs)}
}
\description{
Send emails
}
